Setting up raycast

+++ Raytracing_Renderer/include/color.h
@@ -4,12 +4,12 @@
 using namespace std;
 
 struct Color {
-	uint8_t r;
-	uint8_t g;
-	uint8_t b;
+	int r;
+	int g;
+	int b;
 
 	Color();
-	Color(uint8_t red, uint8_t green, uint8_t blue);
+	Color(int red, int green, int blue);
 
 	string ToString() const;
 };
