+++ Raytracing_Renderer/src/color.cpp
@@ -2,7 +2,7 @@
 
 Color::Color() : r(0), g(0), b(0) {}
 
-Color::Color(uint8_t red, uint8_t green, uint8_t blue)
+Color::Color(int red, int green, int blue)
     : r(red), g(green), b(blue) {}
 
 string Color::ToString() const {
